Merge branches

+++ src/MemoryGraph/graph.cs
@@ -471,7 +471,7 @@ namespace Graphs
             m_writer.Clear();
             m_nodes.Count = 0;
 
-            // Create an undefined node, kind of gross because because SetNode expects to have an entry
+            // Create an undefined node, kind of gross because SetNode expects to have an entry
             // in the m_nodes table, so we make a fake one and then remove it.  
             m_undefinedObjDef = m_writer.GetLabel();
             m_nodes.Add(m_undefinedObjDef);
@@ -576,7 +576,7 @@ namespace Graphs
 
             for (int i = 0; i < nodeCount; i++)
             {
-                m_nodes.Add((StreamLabel)deserializer.ReadInt());
+                m_nodes.Add((StreamLabel)(uint)deserializer.ReadInt());
             }
 
             // Read in the Blob stream.
